requirements.txt
Django==1.5.1
Fabric==1.6.0
Markdown==2.3.1
PIL==1.1.7
South==0.7.6
cssmin==0.1.4
cssselect==0.8
distribute==0.6.31
dj-database-url==0.2.1
django-assets==0.8
django-crispy-forms==1.2.4
django-debug-toolbar==0.9.4
-e git://github.com/stefanfoulis/django-filer.git@75eff8db67ba8e4ad514467cf4b08ad37fe8031e#egg=django_filer-dev
django-filter==0.6
django-floppyforms==1.1
django-mptt==0.5.5
django-polymorphic==0.4.2
django-storages==1.1.8
django-tables2==0.13.0
djangorestframework==2.2.7
-e git+git@github.com:smajda/django-djunk-drawer.git@3ae5d69cfccc04f0a557f389056e11f303c8d10c#egg=djunk_drawer-dev
easy-thumbnails==1.2
gunicorn==0.17.2
ipdb==0.7
ipython==0.13.2
lxml==3.1.2
mdx-smartypants==1.3
mechanize==0.2.5
namedentities==1.301
paramiko==1.10.1
psycopg2==2.5
pyScss==1.1.5
pycrypto==2.6
pyquery==1.2.4
python-dateutil==2.1
python-memcached==1.48
pytz==2013b
readline==6.2.4.1
six==1.3.0
tablib==0.9.11
webassets==0.8
wsgiref==0.1.2
